Mozark_office/bitbucket_mozark/ops-2/availability_docker/deviceAvailability/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
dnspython==2.4.2
idna==3.4
mysql-connector-python==8.1.0
protobuf==4.21.12
pymongo==4.5.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
urllib3==2.0.5
